+++ apps/server/src/modules/common/cities/cities.module.ts
@@ -1,9 +1,9 @@
 import { Module } from "@nestjs/common";
-import { CommonLocationRepository } from "../shared/common-location.repository";
+import { commonLocationProviders } from "../infrastructure/common-location.providers";
 import { CitiesController } from "./cities.controller";
 
 @Module({
   controllers: [CitiesController],
-  providers: [CommonLocationRepository],
+  providers: [...commonLocationProviders],
 })
 export class CitiesModule {}